Reduced test case demonstrating incompatibility with Node "path" module
This explains both issues #2093 & #2094. The reason it was working in your reduced test case and it wasn't working in my environment was because my environment was using Nodes "path" module to generate the strings where as the reduced test case was only using explicit strings.

Every file added in this commit is a demonstration of the bug in action. They were generated by gulp when the should not be.